requirements.txt
openai
google-cloud-vision
pymupdf
streamlit
pillow
python-dotenv
